+++ packages/headless/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @reshaped/headless
 
+## 3.11.0-canary.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @reshaped/utilities@3.11.0-canary.3
+
+## 3.11.0-canary.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @reshaped/utilities@3.11.0-canary.2
+
 ## 3.11.0-canary.1
 
 ### Patch Changes